move types.tsx to new folder
