+++ src/components/providers/auth-provider.tsx
@@ -1,6 +1,6 @@
 import { useLocalUser } from "@/hooks/use-local-user";
 import { delay } from "@/lib/utils";
-import { User } from "@/types";
+import { User } from "@/types/types";
 import { ReactNode, createContext } from "react";
 import { useNavigate } from "react-router-dom";
