+++ src/components/ui/member-details.tsx
@@ -33,7 +33,7 @@ import {
 import { StatusBadge } from "./status-badge";
 import { useWindowSize } from "@/hooks/use-window-size";
 import { UpdateMutation } from "@/hooks/use-table-mutations";
-import { Member } from "@/types";
+import { Member } from "@/types/types";
 import { PersonalTab } from "./personal-tab";
 import { MembershipTab } from "./membership-tab";
 import { NoteTab } from "./note-tab";
