+++ src/components/ui/suspend-popover.tsx
@@ -7,7 +7,7 @@ import { Button } from "./button";
 import { Ban } from "lucide-react";
 import { useTranslation } from "react-i18next";
 import { UseFormReturn } from "react-hook-form";
-import { Member } from "@/types";
+import { Member } from "@/types/types";
 import { getDateMonthsLater, getDateWeekLater } from "@/lib/utils";
 import { SuspendPopoverItem } from "./suspend-popover-item";
