+++ src/hooks/use-local-user.tsx
@@ -1,4 +1,4 @@
-import { User, UserSchema } from "@/types";
+import { User, UserSchema } from "@/types/types";
 import { useEffect, useState } from "react";
 
 const key = "user";
