+++ src/lib/utils.ts
@@ -1,4 +1,4 @@
-import { Member, MemberDTO } from "@/types";
+import { Member, MemberDTO } from "@/types/types";
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
